Implement artifacts upload/download for local backend
Closes #173.
